chenge integer to double

+++ 1weekask/1.5.2.cpp
@@ -2,7 +2,7 @@
 #include<cmath>
 int main()
 {
-	int a,b,c;
+	double a,b,c;
 	std::cin>>a>>b>>c;
 	if(a==0)
 	{
